+++ trunk/dim3MapUtility/Headers/maputility.h
@@ -422,7 +422,7 @@ typedef struct		{
 //
 					
 typedef struct		{
-						short								texture_frame[max_map_scenery_model_texture_frame];
+						int									texture_frame[max_map_scenery_model_texture_frame];
 						float								resize;
 						char								model_name[name_str_len],animation_name[name_str_len];
 						bool								contact_object_on,contact_projectile_on,
